RFL second pass: Split public/internal API
