+++ include/RVLFaceLib/RVLFaceLibInternal.h
@@ -1,0 +1,24 @@
+#ifndef RVL_FACE_LIBRARY_INTERNAL_H
+#define RVL_FACE_LIBRARY_INTERNAL_H
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <RVLFaceLib.h>
+#include <RVLFaceLib/RFLi_Controller.h>
+#include <RVLFaceLib/RFLi_DataUtility.h>
+#include <RVLFaceLib/RFLi_Database.h>
+#include <RVLFaceLib/RFLi_HiddenDatabase.h>
+#include <RVLFaceLib/RFLi_MakeTex.h>
+#include <RVLFaceLib/RFLi_MiddleDatabase.h>
+#include <RVLFaceLib/RFLi_Model.h>
+#include <RVLFaceLib/RFLi_NANDAccess.h>
+#include <RVLFaceLib/RFLi_NANDLoader.h>
+#include <RVLFaceLib/RFLi_System.h>
+#include <RVLFaceLib/RFLi_Texture.h>
+#include <RVLFaceLib/RFLi_Types.h>
+
+#ifdef __cplusplus
+}
+#endif
+#endif
