+++ src/RVLFaceLib/RFL_Icon.c
@@ -1,4 +1,4 @@
-#include <RVLFaceLib.h>
+#include <RVLFaceLib/RVLFaceLibInternal.h>
 
 void RFLSetIconDrawDoneCallback(RFLIconDrawDoneCallback callback) {
     RFLiGetManager()->iconDrawCb = callback;
